7email/requirements.txt
google-cloud-storage==2.7.0
google-auth==2.16.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.70.0
python-dotenv==1.0.0